完成除了“推荐”功能外的所有功能oo化
